src/edit/subtitles.py
from typing import Optional
import ffmpeg
import tempfile
import os

try:
    import whisper
except Exception:  # optional dependency fallback
    whisper = None


def burn_subtitles_karaoke(input_path: str, output_path: str, model: str = "tiny", font_size: int = 48, color: str = "#FFFFFF", stroke_color: str = "#000000", stroke_width: int = 3, margin_bottom: int = 180):
    """
    Transcribe with Whisper (if available) and burn animated karaoke-style subtitles.
    If Whisper is unavailable, this no-ops and just copies the input.
    """
    if whisper is None:
        # pass-through
        ffmpeg.input(input_path).output(output_path, c='copy', movflags='faststart').overwrite_output().run(quiet=True)
        return

    tmpdir = tempfile.mkdtemp()
    srt_path = os.path.join(tmpdir, 'subs.srt')

    model_obj = whisper.load_model(model)
    res = model_obj.transcribe(input_path)

    # Build SRT
    def to_ts(sec: float) -> str:
        ms = int(round(sec * 1000))
        h = ms // 3600000
        ms -= h * 3600000
        m = ms // 60000
        ms -= m * 60000
        s = ms // 1000
        ms -= s * 1000
        return f"{h:02d}:{m:02d}:{s:02d},{ms:03d}"

    with open(srt_path, 'w', encoding='utf-8') as f:
        for i, seg in enumerate(res.get('segments', []), 1):
            text = seg.get('text', '').strip()
            if not text:
                continue
            start, end = float(seg['start']), float(seg['end'])
            f.write(f"{i}\n{to_ts(start)} --> {to_ts(end)}\n{text}\n\n")

    # Burn with styling; animated karaoke effect via drawtext isn't trivial; we keep static but styled
    # For animated effect you'd pre-split per word with timestamps. Here we keep it simple and reliable.
    (
        ffmpeg
        .input(input_path)
        .output(
            output_path,
            vf=f"subtitles='{srt_path}':force_style='Fontsize={font_size},PrimaryColour=&H{color.replace('#','')}&,OutlineColour=&H{stroke_color.replace('#','')}&,Outline={stroke_width},MarginV={margin_bottom}'",
            c:v='libx264',
            c:a='copy',
            movflags='faststart',
        )
        .overwrite_output()
        .run(quiet=True)
    )
