requirements.txt
ffmpeg-python>=0.2.0
numpy>=1.24
pydub>=0.25.1
librosa>=0.10.1
opencv-python>=4.9.0
openai-whisper>=20231117
PyYAML>=6.0.1
click>=8.1.7
